+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.19.0] 
+## Unreleased
+## Added
+- groupsService::protect
+- groupsService::unprotect
+
+## [0.19.0]
 ### Added
 - `getBulk` gets an array of groups by their ids